% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-frontier.R
\name{autoplot.gp_frontier}
\alias{autoplot.gp_frontier}
\alias{autoplot.gp_fit}
\alias{autoplot.gp_grade_path}
\title{Autoplot methods for gradepath frontier objects}
\usage{
\method{autoplot}{gp_frontier}(object, ...)

\method{autoplot}{gp_fit}(object, ...)

\method{autoplot}{gp_grade_path}(object, ...)
}
\arguments{
\item{object}{A \code{gp_frontier}, \code{gp_fit}, or \code{gp_grade_path}.}

\item{...}{Further arguments forwarded to \code{\link[=gp_plot_frontier]{gp_plot_frontier()}}.}
}
\value{
A \code{ggplot} object (class \code{c("gg", "ggplot")}); it renders only when
printed.
}
\description{
ggplot2 \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} methods so \code{autoplot(fr)} (a \code{gp_frontier}),
\code{autoplot(fit)} (a \code{gp_fit}), or \code{autoplot(path)} (a \code{gp_grade_path}) draw the
information-reliability frontier via \code{\link[=gp_plot_frontier]{gp_plot_frontier()}} without attaching
ggplot2.
}
\examples{
fit <- readRDS(system.file("extdata/examples/tiny_fit.rds", package = "gradepath"))
autoplot(fit)   # the information-reliability frontier for a gp_fit

}
\seealso{
\code{\link[=gp_plot_frontier]{gp_plot_frontier()}}, \code{\link[=gp_frontier]{gp_frontier()}}

Other gradepath-plots:
\code{\link[=autoplot.gp_report_card]{autoplot.gp_report_card()}},
\code{\link[=gp_plot_discordance]{gp_plot_discordance()}},
\code{\link[=gp_plot_frontier]{gp_plot_frontier()}},
\code{\link[=gp_plot_posterior_contrast]{gp_plot_posterior_contrast()}},
\code{\link[=gp_plot_report_card]{gp_plot_report_card()}},
\code{\link[=scale_color_gradepath]{scale_color_gradepath()}},
\code{\link[=theme_gradepath]{theme_gradepath()}}
}
\concept{gradepath-plots}
