% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-report-card.R
\name{autoplot.gp_report_card}
\alias{autoplot.gp_report_card}
\title{Autoplot method for a gradepath report card}
\usage{
\method{autoplot}{gp_report_card}(object, ...)
}
\arguments{
\item{object}{A \code{gp_report_card}.}

\item{...}{Further arguments forwarded to \code{\link[=gp_plot_report_card]{gp_plot_report_card()}}.}
}
\value{
A \code{ggplot} object (class \code{c("gg", "ggplot")}); it renders only when
printed.
}
\description{
A thin \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}} method so \code{autoplot(card)} (a \code{gp_report_card})
draws the report-card caterpillar via \code{\link[=gp_plot_report_card]{gp_plot_report_card()}} without
attaching ggplot2. (There is no discordance autoplot method:
\code{autoplot.gp_frontier} already maps to the frontier figure.)
}
\examples{
fit <- readRDS(system.file("extdata/examples/tiny_fit.rds", package = "gradepath"))
autoplot(get_report_card(fit))   # the report-card caterpillar for a gp_fit

}
\seealso{
\code{\link[=gp_plot_report_card]{gp_plot_report_card()}}, \code{\link[=gp_report_card]{gp_report_card()}}

Other gradepath-plots:
\code{\link[=autoplot.gp_frontier]{autoplot.gp_frontier()}},
\code{\link[=gp_plot_discordance]{gp_plot_discordance()}},
\code{\link[=gp_plot_frontier]{gp_plot_frontier()}},
\code{\link[=gp_plot_posterior_contrast]{gp_plot_posterior_contrast()}},
\code{\link[=gp_plot_report_card]{gp_plot_report_card()}},
\code{\link[=scale_color_gradepath]{scale_color_gradepath()}},
\code{\link[=theme_gradepath]{theme_gradepath()}}
}
\concept{gradepath-plots}
