% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-objects-output.R
\name{.gp_estimates_id}
\alias{.gp_estimates_id}
\title{Read unit ids from an eb_estimates-like container}
\usage{
.gp_estimates_id(estimates)
}
\description{
Read unit ids from an eb_estimates-like container
}
\keyword{internal}
