% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-objects-output.R
\name{.gp_estimates_se}
\alias{.gp_estimates_se}
\title{Read unit standard errors from an eb_estimates-like container}
\usage{
.gp_estimates_se(estimates)
}
\description{
Read unit standard errors from an eb_estimates-like container
}
\keyword{internal}
