% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{summary.gp_calibration}
\alias{summary.gp_calibration}
\title{Summarize a \code{gp_calibration} (typed one-row data frame)}
\usage{
\method{summary}{gp_calibration}(object, ...)
}
\arguments{
\item{object}{A \code{gp_calibration}.}

\item{...}{Ignored.}
}
\description{
Summarize a \code{gp_calibration} (typed one-row data frame)
}
